+++ src/app/app.module.ts
@@ -73,6 +73,7 @@ import { AccountComponent } from "./account/account.component";
 import { DatasetsTabComponent } from "./account/additional-components/datasets-tab/datasets-tab.component";
 import { AccessTokensTabComponent } from "./auth/settings/tabs/access-tokens-tab/access-tokens-tab.component";
 import { MatSlideToggleModule } from "@angular/material/slide-toggle";
+import { AutofocusDirective } from "./common/directives/autofocus.directive";
 
 const Services = [
     {
@@ -204,6 +205,7 @@ const MatModules = [
         AdminDashboardComponent,
         AccessTokensTabComponent,
         AccountFlowsTabComponent,
+        AutofocusDirective,
     ],
     imports: [
         AppRoutingModule,